Rearrange folders to separate pages and components
