+++ client/src/App.jsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { LandingPage } from "./components/LandingPage";
-import { LoginUser } from "./components/LoginUser";
-import { RegisterUser } from "./components/RegisterUser";
-import UpdateUser from "./components/UpdateUser";
+import { LandingPage } from "./pages/LandingPage";
+import { LoginUser } from "./pages/LoginUser";
+import { RegisterUser } from "./pages/RegisterUser";
+import UpdateUser from "./pages/UpdateUser";
 import Navbar from "./components/Navbar";
-import ViewUser from "./components/ViewUser";
+import ViewUser from "./pages/ViewUser";
 
 function App() {
   return (
